+++ src/onegov.feriennet/onegov/feriennet/views/invoice.py
@@ -5,8 +5,21 @@ from onegov.feriennet import FeriennetApp, _
 from onegov.feriennet.collections import BillingDetails
 from onegov.feriennet.layout import InvoiceLayout
 from onegov.feriennet.views.shared import all_users
+from onegov.pay import process_payment
 from sortedcontainers import SortedDict
 from stdnum import iban
+
+
+@FeriennetApp.view(
+    model=InvoiceItem,
+    permission=Personal,
+)
+def redirect_to_invoice_view(self, request):
+    return request.redirect(request.link(
+        InvoiceItemCollection(
+            request.app.session(), username=self.username, invoice=self.invoice
+        )
+    ))
 
 
 @FeriennetApp.html(
@@ -49,6 +62,17 @@ def view_my_invoices(self, request):
 
     beneficiary = request.app.org.bank_beneficiary
 
+    payment_provider = request.app.default_payment_provider
+
+    def payment_button(title, price):
+        return request.app.checkout_button(
+            button_label=request.translate(_("Pay Online Now")),
+            title=title,
+            price=payment_provider.adjust_price(price),
+            email=self.username,
+            locale=request.locale
+        )
+
     return {
         'title': title,
         'layout': InvoiceLayout(self, request, title),
@@ -57,5 +81,40 @@ def view_my_invoices(self, request):
         'bills': bills,
         'model': self,
         'account': account,
+        'payment_provider': payment_provider,
+        'payment_button': payment_button,
         'beneficiary': beneficiary
     }
+
+
+@FeriennetApp.view(
+    model=InvoiceItemCollection,
+    template='invoices.pt',
+    permission=Personal,
+    request_method='POST')
+def handle_payment(self, request):
+    provider = request.app.default_payment_provider
+    token = request.params.get('payment_token')
+    period = request.params.get('period')
+
+    q = self.query()
+    q = q.filter(InvoiceItem.invoice == period)
+    q = q.filter(InvoiceItem.paid == False)
+
+    items = tuple(q)
+    bill = BillingDetails(period, items)
+    payment = process_payment('cc', bill.price, provider, token)
+    payment.sync()
+
+    if not payment:
+        request.alert(_("Your payment could not be processed"))
+    else:
+
+        for item in items:
+            item.payment = payment
+            item.paid = True
+            item.source = provider.type
+
+        request.success(_("Your payment has been received. Thank you!"))
+
+    return request.redirect(request.link(self))